# docs/swagger.yaml
basePath: /api
definitions:
...
  http.ErrorResponse:
    properties:
      error:
        example: invalid credentials
        type: string
    type: object
...
  http.LoginRequest:
    properties:
      password:
        example: user
        type: string
      username:
        example: user
        type: string
    type: object
  http.LoginResponse:
    properties:
      access_token:
        example: eyJ...
        type: string
      token_type:
        example: Bearer
        type: string
    type: object
...
  http.RegisterRequest:
    properties:
      password:
        example: secure123
        type: string
      username:
        example: newuser
        type: string
    type: object
  http.SuccessMessage:
    properties:
      message:
        example: updated
        type: string
    type: object
...
  http.UpdateUserRequest:
    properties:
      username:
        example: newUserName
        type: string
    type: object
  http.UserResponse:
    properties:
      role:
        example: user
        type: string
      user_id:
        example: 1
        type: integer
      username:
        example: user
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API для расчёта мощности ЦОД (лабораторная №4)
  title: Datacenter Calculator API
  version: "1.0"
paths:
...
  /users/login:
    post:
      consumes:
      - application/json
      description: Возвращает JWT-токен для авторизации
      parameters:
      - description: Данные для входа
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Аутентификация
      tags:
      - auth
  /users/logout:
    post:
      description: Добавляет токен в blacklist (Redis)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.SuccessMessage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - Bearer: []
      summary: Выход из системы
      tags:
      - auth
  /users/me:
    get:
      description: Возвращает данные авторизованного пользователя
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.UserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - Bearer: []
      summary: Получить профиль текущего пользователя
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Обновляет поля профиля (например, username)
      parameters:
      - description: Поля для обновления
        in: body
        name: request
        schema:
          $ref: '#/definitions/http.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.SuccessMessage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - Bearer: []
      summary: Обновить профиль
      tags:
      - users
  /users/register:
    post:
      consumes:
      - application/json
      description: Создаёт нового пользователя с ролью "user"
      parameters:
      - description: Данные регистрации
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/http.SuccessMessage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Регистрация нового пользователя
      tags:
      - users
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"